requirements.txt
fastapi
uvicorn
# ... existing deps ...
aiogram>=3.0.0
langgraph
langchain
langchain-community
langchain-ollama
chromadb
duckduckgo-search
faster-whisper
python-dotenv
pydantic
aiohttp
